Implements exact match evaluation.
Exact match means that the evaluated sentence is tagged exactly the
same (that is, for all tokens) than the test sentence.

Typical use case is evaluating parsing.

From Dependency Parsing. Kubler, Mcdonald and Nivre (p.79):
     "Exact match: This is the percentage of completely correctly parsed sentences.
The same measure is also used for the evaluation of constituent
parsers."

Our implementation is slightly more general in that it is allowed for
the user to choose the compared fields (so that, for instance, a
tagger or a lemmatizer could be evaluated for exact match as well).